requirements.txt
# Core dependencies for Runaway Guys app
streamlit>=1.28.0
boto3>=1.28.0
pandas>=2.0.0
numpy>=1.24.0

# Development dependencies
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
pytest>=7.4.0

# Optional dependencies (for notebooks and analysis)
matplotlib>=3.7.0
plotnine>=0.12.0
scikit-learn>=1.3.0
jupyterlab>=4.0.0
ipython>=8.12.0
notebook>=7.0.0

# AWS (if needed)
awscli>=1.29.0

# Documentation
mkdocs>=1.5.0

# Utilities
python-dotenv>=1.0.0

# LLM Integration
requests>=2.31.0
groq>=0.4.0
huggingface_hub>=0.20.0

